pyproject.toml
[tool.poetry]
name = "ddd-backend"
version = "0.1.0"
description = "Deep Learning Model served via FastAPI and hosted on Google Cloud Platform"
authors = ["Rohit Nair <6954311+rnair98@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
numpy = ">=1.22,<1.24"
pandas = "^1.5.3"
tensorflow = ">=2.12.0,<3.0.0"
scipy = "^1.10.1"
scikit-learn = "^1.2.2"
mediapipe = "^0.9.2"
opencv-python = "^4.7.0"
matplotlib = "^3.7.1"


[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
google-cloud-storage = "^2.7.0"
autopep8 = "^2.0.2"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
